requirements.txt
docling==2.84.0
ocrmac==1.0.1; sys_platform == "darwin"
onnxruntime==1.24.4
python-docx==1.2.0
python-dotenv==1.2.2
requests==2.33.1
Pillow==12.2.0
